Added Coverart Download
Added a coverart download.  If a coverart is available locally, it will pull from there, else it will attempt to download the coverart and display it.  Corrected error handling some, and have some partially completed code for reading in a rom (this isn't yet ready, so the option is hidden away for now...)