Update to allow searching by title
This update allows searching by partial title.  The program pulls a full copy of the JSON at launch, and the data is seached through on the fly when the name is populated into the search box.  If any results are found, the grid can be clicked on and the details will be populated into the main window.